requirements.txt
numpy
pandas
streamlit
joblib
openpyxl
scikit-learn
Pillow